add modular components for analyzer screen
