+++ lib/screens/poker_analyzer/action_editor_widget.dart
@@ -309,129 +309,6 @@ class _AnalyzeButtonSection extends StatelessWidget {
   }
 }
 
-class _HandHeaderSection extends StatelessWidget {
-  final String handName;
-  final int playerCount;
-  final String streetName;
-  final VoidCallback onEdit;
-  const _HandHeaderSection({
-    required this.handName,
-    required this.playerCount,
-    required this.streetName,
-    required this.onEdit,
-  });
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(AppConstants.padding16),
-      child: Card(
-        color: AppColors.cardBackground,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(AppConstants.radius8)),
-        child: Padding(
-          padding: const EdgeInsets.all(12),
-          child: Row(
-            children: [
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(handName, style: const TextStyle(color: Colors.white, fontSize: 18, fontWeight: FontWeight.bold)),
-                    const SizedBox(height: 4),
-                    Text('$playerCount players • $streetName', style: const TextStyle(color: Colors.white70)),
-                  ],
-                ),
-              ),
-              IconButton(icon: const Icon(Icons.edit, color: Colors.white), onPressed: onEdit),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _PlayerCountSelector extends StatelessWidget {
-  final int numberOfPlayers;
-  final Map<int, String> playerPositions;
-  final Map<int, PlayerType> playerTypes;
-  final ValueChanged<int>? onChanged;
-  final bool disabled;
-  const _PlayerCountSelector({
-    required this.numberOfPlayers,
-    required this.playerPositions,
-    required this.playerTypes,
-    this.onChanged,
-    this.disabled = false,
-  });
-  @override
-  Widget build(BuildContext context) {
-    return DropdownButton<int>(
-      value: numberOfPlayers,
-      dropdownColor: Colors.black,
-      style: const TextStyle(color: Colors.white),
-      iconEnabledColor: Colors.white,
-      items: [for (int i = 2; i <= 10; i++) DropdownMenuItem(value: i, child: Text('Игроков: $i'))],
-      onChanged: disabled ? null : (value) { if (value != null && onChanged != null) onChanged!(value); },
-    );
-  }
-}
-
-class _HandProgressIndicator extends StatelessWidget {
-  final int step;
-  const _HandProgressIndicator({required this.step});
-  @override
-  Widget build(BuildContext context) {
-    const labels = ['Игроки', 'Карты', 'Действия', 'Шоудаун'];
-    const icons = [Icons.people, Icons.style, Icons.list_alt, Icons.flag];
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: List.generate(labels.length, (i) {
-          final active = step >= i;
-          return Expanded(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                AnimatedContainer(
-                  duration: const Duration(milliseconds: 300),
-                  padding: const EdgeInsets.all(6),
-                  decoration: BoxDecoration(
-                    shape: BoxShape.circle,
-                    color: active ? Colors.blueAccent : Colors.white12,
-                  ),
-                  child: Icon(icons[i], size: 16, color: Colors.white),
-                ),
-                const SizedBox(height: 4),
-                Text(labels[i], textAlign: TextAlign.center, style: TextStyle(color: active ? Colors.white : Colors.white54, fontSize: 12)),
-              ],
-            ),
-          );
-        }),
-      ),
-    );
-  }
-}
-
-class _InfoBanner extends StatelessWidget {
-  final String message;
-  const _InfoBanner({required this.message});
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 4.0),
-      child: Container(
-        width: double.infinity,
-        padding: const EdgeInsets.all(12.0),
-        decoration: BoxDecoration(
-          color: AppColors.cardBackground,
-          borderRadius: BorderRadius.circular(AppConstants.radius8),
-        ),
-        child: Text(message, style: const TextStyle(color: Colors.white70)),
-      ),
-    );
-  }
-}
 
 class _HandEditorSection extends StatelessWidget {
   final ActionHistoryService actionHistory;
